src/api/requirements.txt
-i https://pypi.org/simple
annotated-types==0.6.0; python_version >= '3.8'
anyio==4.3.0; python_version >= '3.8'
cachetools==5.3.3; python_version >= '3.7'
certifi==2024.2.2; python_version >= '3.6'
chardet==5.2.0; python_version >= '3.7'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
fastapi==0.110.1; python_version >= '3.8'
google-api-core[grpc]==2.18.0; python_version >= '3.7'
google-auth==2.29.0; python_version >= '3.7'
google-cloud==0.34.0
google-cloud-core==2.4.1; python_version >= '3.7'
google-cloud-discoveryengine==0.11.10; python_version >= '3.7'
google-cloud-storage==2.16.0; python_version >= '3.7'
google-crc32c==1.5.0; python_version >= '3.7'
google-resumable-media==2.7.0; python_version >= '3.7'
googleapis-common-protos[grpc]==1.63.0; python_version >= '3.7'
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==21.2.0; python_version >= '3.5'
h11==0.14.0; python_version >= '3.7'
idna==3.6; python_version >= '3.5'
markdown==3.6; python_version >= '3.8'
packaging==24.0; python_version >= '3.7'
pillow==10.3.0; python_version >= '3.8'
proto-plus==1.23.0; python_version >= '3.6'
protobuf==4.25.3; python_version >= '3.8'
pyasn1==0.6.0; python_version >= '3.8'
pyasn1-modules==0.4.0; python_version >= '3.8'
pydantic==2.6.4; python_version >= '3.8'
pydantic-core==2.16.3; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
python-multipart==0.0.9; python_version >= '3.8'
reportlab==4.1.0; python_version >= '3.7' and python_version < '4'
requests==2.31.0; python_version >= '3.7'
rsa==4.9; python_version >= '3.6' and python_version < '4'
sniffio==1.3.1; python_version >= '3.7'
starlette==0.37.2; python_version >= '3.8'
typing-extensions==4.10.0; python_version >= '3.8'
urllib3==2.2.1; python_version >= '3.8'
uvicorn==0.29.0; python_version >= '3.8'
